requirements.txt
pymongo
optunity
smac
scikit-learn
scipy
matplotlib
networkx
hyperopt
numpy
hpolib2
ConfigSpace
openml
psycopg2
auto-sklearn